+++ pmd-kotlin/src/test/java/net/sourceforge/pmd/lang/kotlin/ast/KotlinParserTests.java
@@ -4,9 +4,18 @@
 
 package net.sourceforge.pmd.lang.kotlin.ast;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.containsString;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
+
+import net.sourceforge.pmd.lang.ast.LexException;
+import net.sourceforge.pmd.lang.ast.ParseException;
+import net.sourceforge.pmd.lang.document.FileId;
 
 /**
  * Miscellaneous Kotlin parser regression tests.
@@ -16,6 +25,39 @@ class KotlinParserTests extends BaseKotlinTreeDumpTest {
     @Test
     void testSimpleKotlin() {
         doTest("Simple");
+    }
+
+    @Test
+    void syntaxErrorThrowsParseException() {
+        ParseException parseException = assertThrows(ParseException.class, () ->
+                KotlinParsingHelper.DEFAULT.parse("fun foo( { }", null, FileId.fromPathLikeString("myfile.kt"))
+        );
+        assertThat(parseException.getMessage(), containsString("myfile.kt"));
+        assertEquals("myfile.kt", parseException.getFileId().getOriginalPath());
+    }
+
+    @Test
+    void lexerErrorThrowsLexException() {
+        // ^ is Java XOR but not a valid Kotlin token - triggers a lexer error
+        assertThrows(LexException.class, () ->
+            KotlinParsingHelper.DEFAULT.parse("fun xor(a: Int, b: Int) = (a ^ b)")
+        );
+    }
+
+    @Test
+    void multipleErrorsSholdBeCollectedAsSuppressedExceptions() {
+        String badCode = "package nl.stokpop\n"
+                + "\n"
+                + "fun xor1(a: Int, b: Int) = (a ^ b)\n"
+                + "fun xor2(a: Int, b: Int) = (a ^ b)\n"
+                + "fun broken( { }";
+        LexException lexException = assertThrows(LexException.class, () -> KotlinParsingHelper.DEFAULT.parse(badCode));
+        Throwable[] suppressed = lexException.getSuppressed();
+        assertEquals(4, suppressed.length);
+        assertInstanceOf(ParseException.class, suppressed[0]);
+        assertInstanceOf(LexException.class, suppressed[1]);
+        assertInstanceOf(ParseException.class, suppressed[2]);
+        assertInstanceOf(ParseException.class, suppressed[3]);
     }
 
     // Regression tests for https://github.com/pmd/pmd/issues/6648